Code_cleanup #2: Move files to respective folders
