+++ src/views/pages/HomePage/Events.js
