+++ src/views/pages/HomePage/Hero.js
