+++ src/views/pages/HomePage/index.js
@@ -5,8 +5,8 @@ import Hero from './Hero';
 import Events from './Events';
 import CTA from './CTA';
 import Footer from '../common/Footer';
-import StandForSection from './StandForView';
-import Team from './Team';
+import StandForSection from './AboutUs/AboutUsView/StandForView';
+import Team from './AboutUs/OurTeam/Team';
 import Promoters from './Promoters';
 import MentorExperience from './MentorExperience';
 import WatchOurVideoView from './WatchVideos';
